+++ examples/static-hbase/outputs.tf
@@ -15,5 +15,6 @@ output "ec2-key" {
 }
 
 output "private-key" {
-  value = tls_private_key.emr_private_key.private_key_pem
+  value     = tls_private_key.emr_private_key.private_key_pem
+  sensitive = true
 }