+++ modules/aws-emr-iam/variables.tf
@@ -68,8 +68,13 @@ variable "tags" {
   default     = {}
 }
 
-variable "abac_tags" {
-  type        = map(string)
-  description = "A map of tags to reference inside the condition in ABAC Statements."
+variable "abac_valid_tags" {
+  type        = map(list(string))
+  description = "Valid tags for maintaining resources when using ABAC IAM Policies with Tag Conditions. Make sure `tags` contain a key value specified here."
   default     = {}
+}
+
+variable "vpc_id" {
+  type        = string
+  description = "VPC ID of the network"
 }