+++ variables.tf
@@ -46,18 +46,6 @@ variable "cluster_name" {
   default     = "TAMR-EMR-Cluster"
 }
 
-variable "tamr_cidrs" {
-  type        = list(string)
-  description = "List of CIDRs for Tamr"
-  default     = []
-}
-
-variable "tamr_sgs" {
-  type        = list(string)
-  description = "Security Groups for the Tamr Instance"
-  default     = []
-}
-
 variable "additional_tags" {
   type        = map(string)
   description = "[DEPRECATED: Use `tags` instead] Additional tags to be attached to the resources created."
@@ -70,16 +58,10 @@ variable "tags" {
   default     = {}
 }
 
-variable "abac_tags" {
-  type        = map(string)
-  description = "Map of tags that will be used inside IAM Policies for restricting EMR Service Role access"
-  default = {}
-}
-
-variable "use_abac_tags" {
-  type = bool
-  description = "Whether the IAM Module will insert Conditions based on Tags into the EMR Policies. Use abac_tags variable for specifying tag keys and values. Setting to true will require your Subnet to have the specified tags."
-  default = false
+variable "abac_valid_tags" {
+  type        = map(list(string))
+  description = "Valid tags for maintaining resources when using ABAC IAM Policies with Tag Conditions. Make sure `tags` contain a key value specified here."
+  default     = {}
 }
 
 variable "emr_service_role_name" {
@@ -159,7 +141,7 @@ variable "master_bid_price" {
 
 variable "master_bid_price_as_percentage_of_on_demand_price" {
   type        = number
-  default     = 1
+  default     = 100
   description = "Bid price as percentage of on-demand price for master instances"
 }
 
@@ -275,6 +257,12 @@ variable "core_timeout_duration_minutes" {
   default     = 10
 }
 
+variable "emr_managed_sg_name" {
+  type        = string
+  description = "Name for the EMR managed security group"
+  default     = "TAMR-EMR-Internal"
+}
+
 variable "emr_managed_master_sg_name" {
   type        = string
   description = "Name for the EMR managed master security group"
@@ -287,22 +275,25 @@ variable "emr_managed_core_sg_name" {
   default     = "TAMR-EMR-Core"
 }
 
-variable "emr_additional_master_sg_name" {
-  type        = string
-  description = "Name for the EMR additional master security group"
-  default     = "TAMR-EMR-Master-Additional"
-}
-
-variable "emr_additional_core_sg_name" {
-  type        = string
-  description = "Name for the EMR additional core security group"
-  default     = "TAMR-EMR-Core-Additional"
-}
-
 variable "emr_service_access_sg_name" {
   type        = string
-  description = "Name for the EMR service access security group"
+  description = "Name for the EMR Service Access security group"
   default     = "TAMR-EMR-Service-Access"
+}
+
+variable "emr_managed_master_sg_ids" {
+  type        = list(string)
+  description = "List of EMR managed master security group ids"
+}
+
+variable "emr_managed_core_sg_ids" {
+  type        = list(string)
+  description = "List of EMR managed core security group ids"
+}
+
+variable "emr_service_access_sg_ids" {
+  type        = list(string)
+  description = "List of EMR service access security group ids"
 }
 
 variable "emr_service_iam_policy_name" {